refactor: rename dir (ToastProvider to Toast)
